SP-4627 Degridder: Apply outer kernels afterwards
It might seem like this doesn't make much of a difference, but this is
slightly more efficient in terms of floating point operations - the
inner-most loop now only executes one multiplication and addition
instead of two multiplications and one addition. To be fair, the extra
multiplication is a real one, so it's fairly cheap, but this is still
a good couple of percent of performance.